pyproject.toml
[project]
name = "repair-dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.5",
    "pillow>=10.4.0",
    "requests>=2.32.4",
    "tqdm>=4.66.5",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
eval = [
    "pillow>=10.4.0",
]
